requirements.txt
anthropic>=0.40.0
tavily-python>=0.5.0
pydantic>=2.0
python-docx>=1.1.0
slack-sdk>=3.30.0
python-dotenv>=1.0.0
fastapi>=0.110.0
uvicorn>=0.29.0
